+++ src/routes/Routes.jsx
@@ -18,6 +18,7 @@ const Messages = lazy( prefetch( () => import( '../components/pages/Messages' ) 
 const Documents = lazy( prefetch( () => import( '../components/pages/Documents' ) ) )
 const Appointments = lazy( prefetch( () => import( '../components/pages/Appointments' ) ) ) 
 const Settings = lazy( prefetch( () => import( '../components/pages/Settings' ) ) )
+const AppSettings = lazy( prefetch( () => import( '../components/pages/AppSettings' ) ) )
 
 export default function Routes() {
 
@@ -58,6 +59,7 @@ export default function Routes() {
             <Route exact path='/profile/documents' element={ <Documents /> } />
             <Route exact path='/profile/appointments' element={ <Appointments /> } />
             <Route exact path='/profile/settings' element={ <Settings /> } />
+            <Route exact path='/profile/app-settings' element={ <AppSettings /> } />
 
         </DOMRoutes>